Extract filters component to its own file
